Updated transformer tests to test objects instead of snapshots.
